+++ db/migrate/20260616150100_add_telefonnummer_to_referees.rb
@@ -1,0 +1,5 @@
+class AddTelefonnummerToReferees < ActiveRecord::Migration[7.1]
+  def change
+    add_column :referees, :telefonnummer, :string
+  end
+end